BookMaker Beta Version.  - Fixed duplicated event handlers in clients.js and matches.js.  - Completed the ticket report.  - Added the number of active matches for a particular sport or tournament.  - Now a manager can modify parameters in the agency.

+++ src/main/java/co/com/bookmaker/business_logic/controller/event/TournamentController.java
@@ -259,7 +259,7 @@ public class TournamentController extends GenericController {
             forward(AnalystController.getJSP(AnalystController.SEARCH_TOURNAMENT));
             return;
         }
-        List<Pair<Tournament, Long>> result;
+        List<Pair<Tournament, Integer>> result;
                 
         FinalUser user = auth.sessionUser(request);
         Agency agency = user.getAgency();
